# src/main/java/com/openapi/banking/api/TransactionApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.19).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.openapi.banking.api;

import com.openapi.banking.model.Account;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;


@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2021-05-30T13:23:07.394Z")

@Validated
@Api(value = "transaction", description = "the transaction API")
@RequestMapping(value = "/v2")
public interface TransactionApi {

    @ApiOperation(value = "Check the Account Balance", nickname = "doFundTransfer", notes = "", response = Account.class, tags={ "Funds Transfer", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Account.class) })
    @RequestMapping(value = "/transaction/{sourceAccount}/{targetAccount}/{amount}",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Account> doFundTransfer(@ApiParam(value = "ID of pet to update",required=true) @PathVariable("sourceAccount") String sourceAccount, @ApiParam(value = "ID of pet to update",required=true) @PathVariable("targetAccount") String targetAccount, @ApiParam(value = "Updated status of the pet",required=true) @PathVariable("amount") Integer amount);

}